requirements.txt
python-dotenv
langchain
langchain-openai
wikipedia
streamlit
langchain-community